+++ src/cli/steadybit-service-definition.ts
@@ -4,6 +4,7 @@ import { Command } from 'commander';
 
 import { deleteServiceDefinition } from '../serviceDefinition/delete';
 import { apply } from '../serviceDefinition/apply';
+import { open } from '../serviceDefinition/open';
 import { init } from '../serviceDefinition/init';
 
 const program = new Command();
@@ -16,6 +17,10 @@ program
   .command('apply <path-to-service-definition-file>')
   .description('Upload a service definition.')
   .action(apply);
+program
+  .command('open <path-to-service-definition-file>')
+  .description('Open the service in the Steadybit UI.')
+  .action(open);
 program
   .command('delete <id-or-path-to-service-definition-file>')
   .description('Deletes a service definition.')