+++ src/serviceDefinition/open.ts
@@ -1,0 +1,9 @@
+import opn from 'open';
+
+import { loadServiceDefinition } from './loading';
+import { baseUrl } from '../api';
+
+export async function open(serviceDefinitionPath: string) {
+  const serviceDefinition = await loadServiceDefinition(serviceDefinitionPath);
+  await opn(`${baseUrl}/service/${serviceDefinition.id}`);
+}