+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
  - Correct supported Node.js version to `>=16`.
  - Send a `User-Agent` header when making Steadybit API calls.
+ - Add a `steadybit service open .steadybit.yml` command.
 
 ## v0.3.0
  - Rename `contract` to `service definition`